requirements.txt
requests
psycopg2-binary
python-dotenv
pandas
tqdm

# chunk
spacy
scispacy
pysbd